Better weights have been loaded. Algorithm has been optimized.
