--- scripts/slow_algorithm_of_style_transfering.py
@@ -1,207 +1,0 @@
-from __future__ import print_function
-
-import matplotlib.pyplot as plt
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-import torch.optim as optim
-import torchvision.models as models
-import torchvision.transforms as transforms
-from PIL import Image
-from pathlib import Path
-
-import requests
-from io import BytesIO
-
-# import warnings
-#
-# warnings.filterwarnings('ignore')
-
-device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-# desired size of the output image
-# use small size if no gpu
-
-
-
-
-def get_image(img_url: str) -> BytesIO:
-    return BytesIO(requests.get(img_url).content)
-
-
-def image_loader(image_url, imsize=256) -> torch.tensor:
-    loader = transforms.Compose([
-        transforms.Resize(imsize),
-        transforms.ToTensor()
-    ])
-    image = Image.open(get_image(image_url))
-    image = loader(image).unsqueeze(0)
-    return image.to(device, torch.float)
-
-
-unloader = transforms.ToPILImage()  # reconvert into PIL image
-
-
-def imshow(tensor: torch.tensor, title=None):
-    image = tensor.cpu().clone()
-    image = image.squeeze(0)
-    image = unloader(image)
-    plt.imshow(image)
-    if title is not None:
-        plt.title(title)
-    plt.pause(0.001)
-
-
-def content_loss(input, target):
-    return F.mse_loss(input, target)
-
-
-def gram_matrix(input):
-    a, b, c, d = input.size()
-
-    features = input.view(a * b, c * d)  # resise F_XL into \hat F_XL
-
-    G = torch.mm(features, features.t())  # compute the gram product
-
-    # we 'normalize' the values of the gram matrix
-    # by dividing by the number of element in each feature maps.
-    return G.div(a * b * c * d)
-
-
-def style_loss(input, target):
-    input = gram_matrix(input)
-    target = gram_matrix(target)
-    return F.mse_loss(target, input)
-
-
-class Normalization(nn.Module):
-    def __init__(self, mean, std):
-        super(Normalization, self).__init__()
-        # .view the mean and std to make them [C x 1 x 1] so that they can
-        # directly work with image Tensor of shape [B x C x H x W].
-        # B is batch size. C is number of channels. H is height and W is width.
-        self.mean = torch.tensor(mean).view(-1, 1, 1)
-        self.std = torch.tensor(std).view(-1, 1, 1)
-
-    def forward(self, img):
-        # normalize img
-        return (img - self.mean) / self.std
-
-
-class vgg(nn.Module):
-    features = [0, 5, 10, 19, 28]
-
-    def __init__(self, cnn: nn.Module, cnn_mean, cnn_std):
-        super().__init__()
-        self.model = nn.Sequential(Normalization(cnn_mean, cnn_std))
-        for layer in range(len(cnn[:self.features[-1] + 1])):
-            self.model.add_module(str(layer), cnn[layer])
-
-    def forward(self, x):
-        features = []
-
-        for layer_num, layer in enumerate(self.model):
-            x = layer(x)
-
-            if layer_num in self.features:  # Если слой находится в списке,
-                features.append(x)  # то запоминаем выход слоя для расчета лосса
-
-        return features
-
-
-def img_to_bytes(img: Image) -> BytesIO:
-    byte_arr = BytesIO()
-
-    img_res = img.resize((512, 512))
-    out_dir = Path('outputs')
-
-    img_res.save(out_dir / 'result.png', format='PNG')
-    img_res.save(byte_arr, format='PNG')
-
-    return byte_arr.getvalue()
-
-
-class StyleTransfer(nn.Module):
-
-    def __init__(self, cnn: nn.Sequential, normalization_mean: torch.tensor, normalization_std: torch.tensor, ):
-        super().__init__()
-        self.cnn = cnn
-        self.normalization_mean = normalization_mean
-        self.normalization_std = normalization_std
-
-    def forward(self, content_img_url: str, style_img_url: str, num_steps=300,
-                style_weight=1000000, content_weight=1) -> BytesIO:
-
-        try:
-            content_img = image_loader(content_img_url)
-            style_img = image_loader(style_img_url)
-            input_img: torch.tensor = content_img.clone()
-            print(content_img.shape, style_img.shape, input_img.shape)
-        except Exception as e:
-            print(e)
-            return "Error"
-
-        model = vgg(self.cnn,
-                    self.normalization_mean,
-                    self.normalization_std)
-
-        input_img.requires_grad_(True)
-        model.requires_grad_(False)
-
-        optimizer = optim.LBFGS([input_img])
-
-        content_features = model(content_img)
-        style_features = model(style_img)
-
-        run = [0]
-        while run[0] <= num_steps:
-
-            def closure():
-
-                with torch.no_grad():
-                    input_img.clamp_(0, 1)
-
-                optimizer.zero_grad()
-                model(input_img)
-                style_score = 0
-                content_score = 0
-
-                gen_features = model(input_img)
-
-                for gen_feature, con_feature, st_feature in zip(
-                        gen_features,
-                        content_features,
-                        style_features
-                ):
-                    content_score += content_loss(con_feature, gen_feature)
-                    style_score += style_loss(st_feature, gen_feature)
-
-                style_score *= style_weight
-                content_score *= content_weight
-
-                loss = style_score + content_score
-                loss.backward()
-
-                run[0] += 1
-                # if run[0] % 100 == 0:
-                #     print("run {}:".format(run))
-                #     print('Style Loss : {:4f} Content Loss: {:4f}'.format(
-                #         style_score.item(), content_score.item()))
-                #     print()
-
-                return style_score + content_score
-
-            optimizer.step(closure)
-
-        imshow(input_img)
-        with torch.no_grad():
-            input_img.clamp_(0, 1)
-
-        res = unloader(input_img.cpu().squeeze(0))
-        return img_to_bytes(res)
-
-
-cnn = models.vgg19(pretrained=True).features.to(device).eval()
-
-cnn_normalization_mean = torch.tensor([0.485, 0.456, 0.406]).to(device)
-cnn_normalization_std = torch.tensor([0.229, 0.224, 0.225]).to(device)
